Feat: 식재료 이미지 엔티티 수정 / 알림 기능 일부 구현 (2)

+++ back/src/main/java/refrigerator/back/ingredient/adapter/in/dto/IngredientRegisterRequestDTO.java
@@ -18,7 +18,7 @@ public class IngredientRegisterRequestDTO {
     private Integer capacity;               // 용량
     private String capacityUnit;            // 용량 단위
     private String storageMethod;           // 보관 방식
-    private Long imageId;                   // 사진 (수정 가능성)
+    private Integer imageId;                // 사진 (수정 가능성)
 }
